+++ photon-model-adapters/vsphere/adapter/src/main/java/com/vmware/photon/controller/model/adapters/vsphere/VSphereEndpointAdapterService.java
@@ -20,6 +20,8 @@ import static com.vmware.photon.controller.model.adapterapi.EndpointConfigReques
 import static com.vmware.xenon.common.Operation.STATUS_CODE_BAD_REQUEST;
 
 import java.net.URI;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
@@ -32,6 +34,7 @@ import com.vmware.photon.controller.model.adapters.vsphere.util.connection.Conne
 import com.vmware.photon.controller.model.adapters.vsphere.util.connection.GetMoRef;
 import com.vmware.photon.controller.model.adapters.vsphere.util.finders.Finder;
 import com.vmware.photon.controller.model.adapters.vsphere.util.finders.FinderException;
+import com.vmware.photon.controller.model.resources.ComputeDescriptionService;
 import com.vmware.photon.controller.model.resources.ComputeDescriptionService.ComputeDescription;
 import com.vmware.photon.controller.model.resources.ComputeService.ComputeState;
 import com.vmware.vim25.InvalidPropertyFaultMsg;
@@ -113,6 +116,12 @@ public class VSphereEndpointAdapterService extends StatelessService {
                 cd.zoneId = r.get(ZONE_KEY).orElse(null);
 
                 cd.environmentName = ComputeDescription.ENVIRONMENT_NAME_ON_PREMISE;
+
+                List<String> children = new ArrayList<>();
+                children.add(ComputeDescriptionService.ComputeDescription.ComputeType.VM_GUEST
+                        .toString());
+                cd.supportedChildren = children;
+
                 cd.instanceAdapterReference = AdapterUriUtil.buildAdapterUri(this.getHost(),
                         VSphereUriPaths.INSTANCE_SERVICE);
                 cd.enumerationAdapterReference = AdapterUriUtil.buildAdapterUri(this.getHost(),